+++ src/main/java/de/samply/tasks/BlazeExporterDelegate.java
@@ -1,7 +1,7 @@
 package de.samply.tasks;
 
 import de.samply.file.bundle.PathsBundle;
-import de.samply.spring.MtbaConst;
+import de.samply.utils.PathsBundleUtils;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
 import org.slf4j.Logger;
@@ -15,7 +15,7 @@ public class BlazeExporterDelegate implements JavaDelegate {
     @Override
     public void execute(DelegateExecution delegateExecution) throws Exception {
         logger.info("Export to Blaze");
-        PathsBundle pathsBundle = (PathsBundle) delegateExecution.getVariable(MtbaConst.PATHS_BUNDLE);
+        PathsBundle pathsBundle = PathsBundleUtils.getPathsBundleVariable(delegateExecution);
         //TODO
 
     }