+++ src/main/java/de/samply/tasks/CBioportalExporterDelegate.java
@@ -1,7 +1,7 @@
 package de.samply.tasks;
 
 import de.samply.file.bundle.PathsBundle;
-import de.samply.spring.MtbaConst;
+import de.samply.utils.PathsBundleUtils;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
 import org.slf4j.Logger;
@@ -16,7 +16,7 @@ public class CBioportalExporterDelegate implements JavaDelegate {
     public void execute(DelegateExecution delegateExecution) throws Exception {
 
         logger.info("Export to cBioPortal");
-        PathsBundle pathsBundle = (PathsBundle) delegateExecution.getVariable(MtbaConst.PATHS_BUNDLE);
+        PathsBundle pathsBundle = PathsBundleUtils.getPathsBundleVariable(delegateExecution);
         //TODO
     }
 }