+++ src/main/java/de/samply/tasks/PseudonymGeneratorDelegate.java
@@ -2,6 +2,7 @@ package de.samply.tasks;
 
 import de.samply.file.bundle.PathsBundle;
 import de.samply.spring.MtbaConst;
+import de.samply.utils.PathsBundleUtils;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
 import org.slf4j.Logger;
@@ -16,7 +17,7 @@ public class PseudonymGeneratorDelegate implements JavaDelegate {
   @Override
   public void execute(DelegateExecution delegateExecution) throws Exception {
     logger.info("Generate pseudonyms");
-    PathsBundle pathsBundle = (PathsBundle) delegateExecution.getVariable(MtbaConst.PATHS_BUNDLE);
+    PathsBundle pathsBundle = PathsBundleUtils.getPathsBundleVariable(delegateExecution);
     //TODO
 
   }