+++ src/main/java/de/samply/utils/PathsBundleUtils.java
@@ -1,0 +1,32 @@
+package de.samply.utils;
+
+
+import de.samply.file.bundle.PathsBundle;
+import de.samply.spring.MtbaConst;
+import java.nio.file.Paths;
+import org.camunda.bpm.engine.delegate.DelegateExecution;
+import org.json.JSONArray;
+
+public class PathsBundleUtils {
+
+  public static JSONArray convert (PathsBundle pathsBundle){
+    return new JSONArray(pathsBundle.getAllPaths());
+  }
+
+  public static PathsBundle convert (JSONArray pathsBundle){
+    PathsBundle result = new PathsBundle();
+    if (pathsBundle != null){
+      pathsBundle.iterator().forEachRemaining(path -> result.addPath(Paths.get((String)path)));
+    }
+    return result;
+  }
+
+  public static void addPathsBundleAsVariable (DelegateExecution delegateExecution, PathsBundle pathsBundle){
+    delegateExecution.setVariable(MtbaConst.PATHS_BUNDLE, convert(pathsBundle));
+  }
+
+  public static PathsBundle getPathsBundleVariable (DelegateExecution delegateExecution){
+    return convert ((JSONArray) delegateExecution.getVariable(MtbaConst.PATHS_BUNDLE));
+  }
+
+}